Provision Istio with the operator
Instead of manually installing, upgrading, and uninstalling Istio in a
production environment, the installation can instead be managed using
the Istio operator. This relieves the burden of managing different
`istioctl` versions. Simply update the operator custom resource (CR) and
the operator controller will apply the corresponding configuration
changes.

The same `IstioOperator` API is used to install Istio with the operator
as when using the `istioctl` install instructions. In both cases,
configuration is validated against a schema and the same correctness
checks are performed.

This commit includes an Istio configuration that provisions:

 - An ingress gateway for north-south traffic. This is a standard
   ingress installation.

 - An ingress gateway for east-west traffic. This is configured to
   enable mesh expansion.

 - An outbound traffic policy allowing any request to exit the mesh

 - A default pod disruption budget